+++ src/index.ts
@@ -2,5 +2,6 @@ export * from './add';
 export * from './divide';
 export * from './mean';
 export * from './meanBy';
+export * from './multiply';
 export * from './sum';
 export * from './sumBy';