+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `divide()` function
 - `mean()` function
 - `meanBy()` function
+- `multiply()` function
 - remove lodash-es dependency
 
 ## 1.1.0 - 2022-09-14